Entites are now created and linked on subsequent frames
Previously, if an entity was created during the middle of a frame
( which they always are ) then the iterator being used for the
components currently being updated may become invalid
if the vector resizes itself.

Now, whenever a new entity is requested from lua, it will be created
on the very next frame

If an entity requests to observe the to-be created entities events,
then that's fine too. Event Observers are no longer handled immedately
upon request. They are now processed at the start
of each frame after new entities are instantiated.