Improved API for sending STL containers to LUA
Stack specializations are now written for vector, list, and map
containers. If one is sent to lua, then a new lua table is generated
from the container. Likewise, when Lua calls a function that requires an
stl container and passes a lua table, a container is created from the
contents of the table

Also began breaking up LuaInterface functionality